+++ packages/sitebuilder/src/wizards/go-live/data/constants.ts
@@ -34,9 +34,9 @@ export const GoLiveStringData = {
 	},
 	updateSiteUrl: {
 		screenTitle: __('You\'re ready to go live with', 'moderntribe-sitebuilder'),
-		screenDescription: __('Once we begin, this process may take a few minutes. When the process is complete, you will need to log in again on the new live domain', 'moderntribe-sitebuilder'),
+		screenDescription: __('Once we begin, this process may take a few minutes. When the process is complete, you will need to log in again on the new live domain.', 'moderntribe-sitebuilder'),
+		descriptionProcessing: __('This process will take a few minutes. You\'ll be logged out when it completes and will need to log in again on your new domain.', 'moderntribe-sitebuilder'),
 		launchImgAltText: __('Go live rocket', 'moderntribe-sitebuilder'),
-		descriptionProcessing: __('This can take a few minutes. You can log out now, or we can will just log out upon completion.', 'moderntribe-sitebuilder'),
 		loginUrlLabelText: __('Copy Your New Login URL', 'moderntribe-sitebuilder'),
 		loginUrlHelperText: __('Copy and save this URL. Once your domain is connected you\'ll need it to login.', 'moderntribe-sitebuilder')
 	},
@@ -46,7 +46,7 @@ export const GoLiveStringData = {
 	},
 	claimYourDomain: {
 		screenTitle: __('is all yours!', 'moderntribe-sitebuilder'),
-		screenDescription: __('While a new domain is often ready for use within an hour, it can take up to 8 hours to completely process.', 'moderntribe-sitebuilder'),
+		screenDescription: __('We need a little time to update the entire internet about your purchase. You\'ll be able to connect your domain when we\'re done. This typically takes an hour but can take up to 8 hours to complete.', 'moderntribe-sitebuilder'),
 	},
 	errorDomainGeneral: {
 		accountContent: __('It looks like we\'re having trouble connecting your domain.', 'moderntribe-sitebuilder'),