+++ packages/sitebuilder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @moderntribe/sitebuilder
 
+## 1.6.3
+
+### Patch Changes
+
+- 15fb8ef: Content change for Go Live ClaimYourDomain.tsx
+
+## 1.6.2
+
+### Patch Changes
+
+- 5682a8d: Replaces text from Go Live wizard screen
+
 ## 1.6.1
 
 ### Patch Changes